+++ src/shared/Communication.cpp
@@ -157,6 +157,6 @@ void Communication::Pipe::WriteMessage(Communication::OutputBuffer& message)
 {
   DWORD bytesWritten;
   std::string data = message.Get();
-  if (!WriteFile(pipe, data.c_str(), data.length(), &bytesWritten, 0))
+  if (!WriteFile(pipe, data.c_str(), static_cast<DWORD>(data.length()), &bytesWritten, 0))
     throw std::runtime_error("Failed to write to pipe");
 }
